+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # act-aaa CHANGE LOG
 
 1.5.4
+* update to act-1.8.18
+* Support specifying profile in `aaa.authenticate.list` file #35
 * User cache could not evict when logout #34
 
 1.5.3 09/Dec/2018